// gym-track-angular/src/app/services/models/exercise-response.ts
/* tslint:disable */
/* eslint-disable */
export interface ExerciseResponse {
  id?: number;
  name?: string;
  type?: 'CHEST' | 'SHOULDERS' | 'BICEPS' | 'TRICEPS' | 'FOREARM' | 'TRAPS' | 'LATS' | 'MIDDLE_BACK' | 'LOW_BACK' | 'ABS' | 'QUADRICEPS' | 'GLUTES' | 'ABDUCTORS' | 'ADDUCTORS' | 'HAMSTRINGS' | 'CALF';
}